Multi-line text

+++ Source/ASMBuilder.cpp
@@ -19,7 +19,7 @@ AsmBuilder& AsmBuilder::rel(uint32_t referenceAddress)
 {
 	// Global references are relative to the address of the next operation. Here, we assume that the next operation
 	//   is always 4 bytes after the given address.
-	uint32_t currentAddress = rootAddress + (builder.size() + 4);
+	uint32_t currentAddress = rootAddress + ((uint32_t)builder.size() + 4);
 	uint32_t relativeAddress = referenceAddress - currentAddress;
 	uint8_t* bytes = static_cast<uint8_t*>(static_cast<void*>(&relativeAddress));
 	builder.insert(builder.end(), bytes, bytes + 4);
@@ -54,7 +54,7 @@ const uint8_t* AsmBuilder::get() const
 
 const int AsmBuilder::size() const
 {
-	return builder.size();
+	return (int)builder.size();
 }
 
 void AsmBuilder::printBytes() const
